+++ src/seeders/e-ticket-seeder.ts
@@ -10,9 +10,15 @@ const generateETicket = (
   ticketNumber: faker.number.int({ min: 1, max: 999999 }),
 });
 
-export const seedETickets = async (prisma: PrismaClient, count = 25) => {
+export const seedETickets = async (
+  prisma: PrismaClient,
+  count = 25,
+  deleteFirst = true
+) => {
   console.log("🌱 Seeding ETickets...");
-  await prisma.eTicket.deleteMany();
+  if (deleteFirst) {
+    await prisma.eTicket.deleteMany();
+  }
   const users = await prisma.user.findMany({ select: { id: true } });
   const orders = await prisma.order.findMany({ select: { id: true } });
   if (!users.length || !orders.length) {
